+++ libs/base_num/src/float/mod.rs
@@ -3,18 +3,25 @@
 #![doc = crate::_DOC_NUM_FLOAT!()]
 //
 
+pub(crate) mod _consts; // PI, TAU, SQRT2, …
 pub(crate) mod _docs; // _FLOAT_[ALGORITHM|FORMULA|NOTATION|PIECEWISE]_*!()
-// mod float_const; // FloatConst
+mod _internals; // _FloatInternals
+
+mod float_const; // FloatConst
 // mod wrapper; // Float
 
 crate::structural_mods! { // _mods
     _mods {
-        // pub use super::{
-        //     float_const::_all::*,
-        //     wrapper::_all::*,
-        // };
+        pub use super::{
+            float_const::*,
+            // wrapper::_all::*,
+        };
     }
     _workspace_internals {
-        pub use super::_docs::*;
+        pub use super::{
+            _consts::*,
+            _docs::*,
+            _internals::*,
+        };
     }
 }