+++ src/num/float/mod.rs
@@ -4,23 +4,23 @@
 //
 
 mod ext_float; // FloatExt
-mod float_const; // FloatConst
 mod wrapper; // Float
 
+// re-exports
 crate::mod_path!(_c "../../../libs/base_core/src/num/float/reexports.rs");
 
-crate::structural_mods! { // _mods, crate_internals
+crate::structural_mods! { // _mods
     _mods {
         pub use super::{
             ext_float::*,
-            float_const::*,
             wrapper::*,
         };
 
-        /* re-exports */
+        // re-exports
         pub use super::_c::*;
-        // aliases
         #[doc(inline)]
         pub use devela_base_core::{f32bits, f32bits_niche, f64bits, f64bits_niche, fsize};
+        #[doc(inline)]
+        pub use devela_base_num::{FloatConst};
     }
 }